Thanks Chanmann

+++ nanopay/src/net/nanopay/meter/report/PaymentSummaryReportDAO.js
@@ -232,7 +232,7 @@ foam.CLASS({
       javaCode: `
         NumTotal nt = new NumTotal();
         nt.num = ((Count) transactions.inX(x).select(new Count())).getValue();
-        nt.total = Math.round((Double) ((Sum) transactions.inX(x).select(SUM(Transaction.AMOUNT))).getValue());
+        nt.total = Math.round(((Sum) transactions.inX(x).select(SUM(Transaction.AMOUNT))).getValue());
         return nt;
       `
     },
